verify specified tubes exist before executing for certain commands (#14)
Problem

When running `purge_buried` against a tube, the output indicated 0 jobs purged, even though beanmon showed 1 buried job. After a lot of digging around realized that I had incorrectly prepended the tube name with the cluster name, but the tooling never told me that was wrong. (The separate `stats_tubes` command did raise a "tube not found" error, but the prior of `purge_buried` showing 0 and not raising that exception biased me towards thinking something was wrong with the beanmon infrastructure, or the beancmd utility).

Solution

For certain commands, we first verify that the given tubes actually exist before running the logic against them.

I picked the commands where it would be surprising to "see a 0":

* `bury`
* `flush`
* `kick`
* `pause`
* `purge_buried`

Example error message:
```
$ venv/bin/python -m beancmd.beancmd purge_buried -H localhost fdf
Could not run command: Unable to find requested tubes (['fdf']) in server tubes (['default'])
```

Because `beancmd` has not been updated in a while this PR also:

* fixes up a help description.
* changes the `pystalk` client version to match other projects
* introduces a `Makefile` similar to other python projects.
* because `beancmd` is packaged to use `python-3.9`, `ci.yml` now only uses `3.9`.
* `ci.yml` now delegates creating the venv, installing deps, and running tests to the `make test` target.